connect-and-join: extract client component into /modules folder
