+++ apps/web/app/future/connect-and-join/page.tsx
@@ -1,4 +1,5 @@
-import LegacyPage from "@pages/connect-and-join";
 import { WithLayout } from "app/layoutHOC";
+
+import LegacyPage from "~/connect-and-join/connect-and-join-view";
 
 export default WithLayout({ getLayout: null, Page: LegacyPage })<"P">;
