+++ apps/web/pages/more.tsx
@@ -4,3 +4,5 @@ import MorePage from "~/more/more-page-view";
 
 const Page = () => <MorePage />;
 Page.PageWrapper = PageWrapper;
+
+export default Page;
