+++ apps/web/pages/upgrade.tsx
@@ -4,3 +4,5 @@ import UpgradePage from "~/upgrade/upgrade-view";
 
 const Page = () => <UpgradePage />;
 Page.PageWrapper = PageWrapper;
+
+export default Page;
